Auto stash before checking out "HEAD"
改动了蓝牙串口，改动映射函数，加上差值检测

+++ BSP/pid.c
@@ -23,8 +23,8 @@ void abs_limit(float *a, float ABS_MAX)
 static void pid_param_init(
     pid_t *pid,
     uint32_t mode,
-    uint32_t maxout,
-    uint32_t intergral_limit,
+    float maxout,
+    float intergral_limit,
     float kp,
     float ki,
     float kd)
@@ -157,8 +157,8 @@ float pid_sp_calc(pid_t *pid, float get, float set, float gyro)
 void PID_struct_init(
     pid_t *pid,
     uint32_t mode,
-    uint32_t maxout,
-    uint32_t intergral_limit,
+    float maxout,
+    float intergral_limit,
 
     float kp,
     float ki,
